requirements.txt
as==0.1
astroid==2.4.2
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
idna==2.8
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
pylint==2.6.0
PySocks==1.7.1
python-decouple==3.3
pyttsx3==2.71
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
silence-tensorflow==1.1.1
simplejson==3.17.2
six==1.15.0
soupsieve==1.9.5
tf==1.0.0
tflearn==0.3.2
toml==0.10.1
tqdm==4.50.0
tweepy==3.9.0
typed-ast==1.4.1
urllib3==1.25.7
Werkzeug==1.0.1
wikipedia==1.4.0

